+++ pext/qml/main.qml
@@ -610,6 +610,7 @@ ApplicationWindow {
                 }
 
                 MenuItem {
+                    visible: platform != 'Darwin'
                     objectName: "menuOutputAutoType"
                     text: qsTr("Type automatically")
                     checkable: true
@@ -715,7 +716,6 @@ ApplicationWindow {
                 text: qsTr("Always show tray icon")
                 checkable: true
             }
-
         }
 
         Menu {
@@ -734,6 +734,12 @@ ApplicationWindow {
             MenuItem {
                 objectName: "menuCheckForUpdates"
                 text: qsTr("Check for updates")
+            }
+
+            MenuItem {
+                visible: platform == 'Darwin'
+                objectName: "menuInstallQuickActionService"
+                text: qsTr("Install quick action service")
             }
 
             MenuItem {